refactor: split home auth and signin boundaries
