+++ dogArea/Source/UserdefaultSetting.swift
@@ -1106,775 +1106,3 @@ final class SyncOutboxStore {
 }
 
 
-enum AppSessionState: Equatable {
-    case guest
-    case member(userId: String)
-
-    var isMember: Bool {
-        if case .member = self { return true }
-        return false
-    }
-}
-
-enum FeatureCapability: String, CaseIterable {
-    case walkRead = "walk_read"
-    case walkWrite = "walk_write"
-    case cloudSync = "cloud_sync"
-    case aiGeneration = "ai_generation"
-    case nearbySocial = "nearby_social"
-}
-
-enum FeatureGateDecision: Equatable {
-    case allowed
-    case requiresMember(MemberUpgradeTrigger)
-}
-
-enum AppFeatureGate {
-    private enum AccessPolicy {
-        case guestAllowed
-        case memberOnly(trigger: MemberUpgradeTrigger)
-    }
-
-    private static let matrix: [FeatureCapability: AccessPolicy] = [
-        .walkRead: .guestAllowed,
-        .walkWrite: .guestAllowed,
-        .cloudSync: .memberOnly(trigger: .walkHistory),
-        .aiGeneration: .memberOnly(trigger: .imageGenerator),
-        .nearbySocial: .memberOnly(trigger: .walkHistory),
-    ]
-
-    static func currentSession() -> AppSessionState {
-        guard let identity = DefaultAuthSessionStore.shared.currentIdentity(),
-              identity.userId.isEmpty == false else {
-            return .guest
-        }
-        guard DefaultAuthSessionStore.shared.currentTokenSession() != nil else {
-            return .guest
-        }
-        return .member(userId: identity.userId)
-    }
-
-    static func decision(for capability: FeatureCapability, session: AppSessionState) -> FeatureGateDecision {
-        guard let policy = matrix[capability] else {
-            return .allowed
-        }
-        switch policy {
-        case .guestAllowed:
-            return .allowed
-        case .memberOnly(let trigger):
-            return session.isMember ? .allowed : .requiresMember(trigger)
-        }
-    }
-
-    static func isAllowed(_ capability: FeatureCapability, session: AppSessionState = currentSession()) -> Bool {
-        if case .allowed = decision(for: capability, session: session) {
-            return true
-        }
-        return false
-    }
-}
-
-enum MemberUpgradeTrigger: String {
-    case walkStart = "walk_start"
-    case imageGenerator = "image_generator"
-    case walkHistory = "walk_history"
-    case walkBackup = "walk_backup"
-
-    var title: String {
-        switch self {
-        case .walkStart:
-            return "회원 전환 후 산책 기록"
-        case .imageGenerator:
-            return "회원 전환 후 이미지 생성"
-        case .walkHistory:
-            return "회원 전환 후 기록 동기화"
-        case .walkBackup:
-            return "로그인하고 산책 백업"
-        }
-    }
-
-    var message: String {
-        switch self {
-        case .walkStart:
-            return "산책 기록은 계정과 연결되어야 안전하게 저장되고 기기 간 동기화됩니다."
-        case .imageGenerator:
-            return "AI 이미지 생성 결과를 안정적으로 저장하려면 계정 연동이 필요합니다."
-        case .walkHistory:
-            return "지금 로그인하면 현재 기기의 산책 기록을 계정에 백업하고 다른 기기에서도 볼 수 있어요."
-        case .walkBackup:
-            return "게스트 모드 기록은 기기 삭제 시 유실될 수 있어요. 로그인 후 자동 백업을 켜세요."
-        }
-    }
-}
-
-struct MemberUpgradeRequest: Identifiable {
-    let id = UUID()
-    let trigger: MemberUpgradeTrigger
-}
-
-struct GuestDataUpgradeSnapshot: Equatable {
-    let sessionCount: Int
-    let pointCount: Int
-    let totalAreaM2: Double
-    let totalDurationSec: Double
-    let sessionIds: [String]
-    let signature: String
-}
-
-struct GuestDataUpgradeReport: Codable, Equatable, Identifiable {
-    var id: String { userId + ":" + signature }
-    let userId: String
-    let signature: String
-    let sessionCount: Int
-    let pointCount: Int
-    let totalAreaM2: Double
-    let totalDurationSec: Double
-    let pendingCount: Int
-    let permanentFailureCount: Int
-    let lastErrorCode: String?
-    let remoteSessionCount: Int?
-    let remotePointCount: Int?
-    let remoteTotalAreaM2: Double?
-    let remoteTotalDurationSec: Double?
-    let validationPassed: Bool?
-    let validationMessage: String?
-    let executedAt: TimeInterval
-
-    var hasOutstandingWork: Bool {
-        pendingCount > 0 || permanentFailureCount > 0
-    }
-}
-
-struct GuestDataUpgradePrompt: Identifiable {
-    let id = UUID()
-    let snapshot: GuestDataUpgradeSnapshot
-    let shouldEmphasizeRetry: Bool
-}
-
-final class GuestDataUpgradeService {
-    static let shared = GuestDataUpgradeService()
-
-    private struct ReportSeed {
-        let signature: String
-        let sessionCount: Int
-        let pointCount: Int
-        let totalAreaM2: Double
-        let totalDurationSec: Double
-    }
-
-    private let syncOutbox = SyncOutboxStore.shared
-    private let syncTransport = SupabaseSyncOutboxTransport()
-    private let walkRepository: WalkRepositoryProtocol
-    private let reportStoragePrefix = "guest.data.upgrade.report.v1."
-    private let acknowledgedSignaturePrefix = "guest.data.upgrade.signature.v1."
-
-    private init(walkRepository: WalkRepositoryProtocol = WalkRepositoryContainer.shared) {
-        self.walkRepository = walkRepository
-    }
-
-    func pendingPrompt(for userId: String) -> GuestDataUpgradePrompt? {
-        guard let snapshot = localSnapshot(), snapshot.sessionCount > 0 else { return nil }
-        let report = latestReport(for: userId)
-        let acknowledgedSignature = UserDefaults.standard.string(forKey: signatureKey(for: userId))
-        if acknowledgedSignature == snapshot.signature, report?.hasOutstandingWork == false {
-            return nil
-        }
-        return GuestDataUpgradePrompt(
-            snapshot: snapshot,
-            shouldEmphasizeRetry: report?.hasOutstandingWork == true
-        )
-    }
-
-    func latestReport(for userId: String) -> GuestDataUpgradeReport? {
-        guard let data = UserDefaults.standard.data(forKey: reportKey(for: userId)),
-              let decoded = try? JSONDecoder().decode(GuestDataUpgradeReport.self, from: data) else {
-            #if DEBUG
-            print("[GuestUpgrade] latestReport: no cached report user=\(userId)")
-            #endif
-            return nil
-        }
-        #if DEBUG
-        print(
-            "[GuestUpgrade] latestReport: user=\(userId) pending=\(decoded.pendingCount) permanent=\(decoded.permanentFailureCount) lastError=\(decoded.lastErrorCode ?? "none")"
-        )
-        #endif
-        return decoded
-    }
-
-    func runUpgrade(for userId: String, forceRetry: Bool = false) async -> GuestDataUpgradeReport? {
-        let snapshot = localSnapshot()
-        let previousReport = latestReport(for: userId)
-
-        if forceRetry {
-            if let snapshot, snapshot.sessionIds.isEmpty == false {
-                syncOutbox.requeuePermanentFailures(walkSessionIds: Set(snapshot.sessionIds))
-                #if DEBUG
-                print("[GuestUpgrade] requeue permanent failures for \(snapshot.sessionIds.count) local sessions")
-                #endif
-            } else {
-                syncOutbox.requeuePermanentFailures()
-                #if DEBUG
-                print("[GuestUpgrade] requeue permanent failures for all outbox sessions")
-                #endif
-            }
-        }
-
-        if let snapshot, snapshot.sessionCount > 0 {
-            #if DEBUG
-            print(
-                "[GuestUpgrade] runUpgrade start user=\(userId) forceRetry=\(forceRetry) sessions=\(snapshot.sessionCount) points=\(snapshot.pointCount)"
-            )
-            #endif
-            var enqueuedSessionCount = 0
-            for polygon in walkRepository.fetchPolygons() {
-                guard let sessionDTO = WalkBackfillDTOConverter.makeSessionDTO(
-                    from: polygon,
-                    ownerUserId: userId,
-                    petId: nil,
-                    sourceDevice: "ios"
-                ) else { continue }
-                syncOutbox.enqueueWalkStages(sessionDTO: sessionDTO)
-                enqueuedSessionCount += 1
-            }
-            #if DEBUG
-            print("[GuestUpgrade] enqueue completed: \(enqueuedSessionCount) sessions queued")
-            #endif
-        } else {
-            #if DEBUG
-            print("[GuestUpgrade] runUpgrade continue without local sessions user=\(userId)")
-            #endif
-        }
-
-        let summary = await syncOutbox.flush(using: syncTransport, now: Date())
-        #if DEBUG
-        print(
-            "[GuestUpgrade] flush summary pending=\(summary.pendingCount) permanent=\(summary.permanentFailureCount) lastError=\(summary.lastErrorCode?.rawValue ?? "none")"
-        )
-        #endif
-
-        let hasOutstanding = summary.pendingCount > 0 || summary.permanentFailureCount > 0
-        guard hasOutstanding || snapshot != nil || previousReport != nil else {
-            clearPersistedReport(for: userId)
-            #if DEBUG
-            print("[GuestUpgrade] runUpgrade done: no local sessions and no outbox work user=\(userId)")
-            #endif
-            return nil
-        }
-
-        let seed = reportSeed(snapshot: snapshot, fallback: previousReport)
-        var remoteSummary: SyncBackfillValidationSummary? = nil
-        var validation: (passed: Bool?, message: String?) = (nil, hasOutstanding ? "local_snapshot_unavailable" : nil)
-
-        if let snapshot {
-            remoteSummary = await syncTransport.fetchBackfillValidationSummary(sessionIds: snapshot.sessionIds)
-            #if DEBUG
-            if let remoteSummary {
-                print(
-                    "[GuestUpgrade] remote summary sessions=\(remoteSummary.sessionCount) points=\(remoteSummary.pointCount) area=\(remoteSummary.totalAreaM2) duration=\(remoteSummary.totalDurationSec)"
-                )
-            } else {
-                print("[GuestUpgrade] remote summary unavailable")
-            }
-            #endif
-            validation = validate(local: snapshot, remote: remoteSummary)
-        }
-
-        let report = GuestDataUpgradeReport(
-            userId: userId,
-            signature: seed.signature,
-            sessionCount: seed.sessionCount,
-            pointCount: seed.pointCount,
-            totalAreaM2: seed.totalAreaM2,
-            totalDurationSec: seed.totalDurationSec,
-            pendingCount: summary.pendingCount,
-            permanentFailureCount: summary.permanentFailureCount,
-            lastErrorCode: summary.lastErrorCode?.rawValue,
-            remoteSessionCount: remoteSummary?.sessionCount ?? previousReport?.remoteSessionCount,
-            remotePointCount: remoteSummary?.pointCount ?? previousReport?.remotePointCount,
-            remoteTotalAreaM2: remoteSummary?.totalAreaM2 ?? previousReport?.remoteTotalAreaM2,
-            remoteTotalDurationSec: remoteSummary?.totalDurationSec ?? previousReport?.remoteTotalDurationSec,
-            validationPassed: validation.passed,
-            validationMessage: validation.message,
-            executedAt: Date().timeIntervalSince1970
-        )
-
-        persist(report: report, for: userId)
-        if report.hasOutstandingWork == false {
-            UserDefaults.standard.set(seed.signature, forKey: signatureKey(for: userId))
-            if seed.sessionCount == 0, seed.pointCount == 0 {
-                clearPersistedReport(for: userId)
-            }
-        }
-        #if DEBUG
-        let validationText: String = {
-            guard let passed = report.validationPassed else { return "nil" }
-            return passed ? "true" : "false"
-        }()
-        print(
-            "[GuestUpgrade] runUpgrade done user=\(userId) outstanding=\(report.hasOutstandingWork) validation=\(validationText) message=\(report.validationMessage ?? "none")"
-        )
-        #endif
-        return report
-    }
-
-    private func localSnapshot() -> GuestDataUpgradeSnapshot? {
-        let polygons = walkRepository.fetchPolygons()
-        guard polygons.isEmpty == false else { return nil }
-
-        let sessionIds = polygons.map { $0.id.uuidString.lowercased() }.sorted()
-        let pointCount = polygons.reduce(0) { $0 + $1.locations.count }
-        let totalArea = polygons.reduce(0.0) { $0 + $1.walkingArea }
-        let totalDuration = polygons.reduce(0.0) { $0 + $1.walkingTime }
-        let signature = signatureForSnapshot(
-            sessionIds: sessionIds,
-            pointCount: pointCount,
-            totalAreaM2: totalArea,
-            totalDurationSec: totalDuration
-        )
-        return GuestDataUpgradeSnapshot(
-            sessionCount: sessionIds.count,
-            pointCount: pointCount,
-            totalAreaM2: totalArea,
-            totalDurationSec: totalDuration,
-            sessionIds: sessionIds,
-            signature: signature
-        )
-    }
-
-    /// 게스트 이관 리포트 작성을 위한 기준 값을 계산합니다.
-    /// - Parameters:
-    ///   - snapshot: 현재 로컬 산책 스냅샷입니다.
-    ///   - fallback: 기존에 저장된 마지막 리포트입니다.
-    /// - Returns: 로컬 스냅샷이 있으면 우선 사용하고, 없으면 기존 리포트 기반으로 복원한 기준 값입니다.
-    private func reportSeed(
-        snapshot: GuestDataUpgradeSnapshot?,
-        fallback: GuestDataUpgradeReport?
-    ) -> ReportSeed {
-        if let snapshot {
-            return ReportSeed(
-                signature: snapshot.signature,
-                sessionCount: snapshot.sessionCount,
-                pointCount: snapshot.pointCount,
-                totalAreaM2: snapshot.totalAreaM2,
-                totalDurationSec: snapshot.totalDurationSec
-            )
-        }
-        return ReportSeed(
-            signature: fallback?.signature ?? "outbox-only",
-            sessionCount: fallback?.sessionCount ?? 0,
-            pointCount: fallback?.pointCount ?? 0,
-            totalAreaM2: fallback?.totalAreaM2 ?? 0,
-            totalDurationSec: fallback?.totalDurationSec ?? 0
-        )
-    }
-
-    private func persist(report: GuestDataUpgradeReport, for userId: String) {
-        guard let data = try? JSONEncoder().encode(report) else { return }
-        UserDefaults.standard.set(data, forKey: reportKey(for: userId))
-        #if DEBUG
-        print("[GuestUpgrade] report persisted user=\(userId) key=\(reportKey(for: userId))")
-        #endif
-    }
-
-    /// 저장된 게스트 이관 리포트를 제거해 stale 카드 노출을 방지합니다.
-    /// - Parameter userId: 리포트를 삭제할 멤버 사용자 식별자입니다.
-    /// - Returns: 없음. 지정 사용자의 로컬 리포트 캐시를 삭제합니다.
-    private func clearPersistedReport(for userId: String) {
-        UserDefaults.standard.removeObject(forKey: reportKey(for: userId))
-        #if DEBUG
-        print("[GuestUpgrade] report cleared user=\(userId) key=\(reportKey(for: userId))")
-        #endif
-    }
-
-    private func validate(
-        local: GuestDataUpgradeSnapshot,
-        remote: SyncBackfillValidationSummary?
-    ) -> (passed: Bool?, message: String?) {
-        guard let remote else {
-            return (nil, "remote_summary_unavailable")
-        }
-        let areaTolerance = max(1.0, local.totalAreaM2 * 0.01)
-        let durationTolerance = max(3.0, local.totalDurationSec * 0.01)
-
-        let sessionMatched = local.sessionCount == remote.sessionCount
-        let pointMatched = local.pointCount == remote.pointCount
-        let areaMatched = abs(local.totalAreaM2 - remote.totalAreaM2) <= areaTolerance
-        let durationMatched = abs(local.totalDurationSec - remote.totalDurationSec) <= durationTolerance
-        let passed = sessionMatched && pointMatched && areaMatched && durationMatched
-
-        let message = passed
-        ? "validated"
-        : [
-            sessionMatched ? nil : "session_mismatch",
-            pointMatched ? nil : "point_mismatch",
-            areaMatched ? nil : "area_mismatch",
-            durationMatched ? nil : "duration_mismatch"
-        ].compactMap { $0 }.joined(separator: ",")
-
-        return (passed, message)
-    }
-
-    private func reportKey(for userId: String) -> String {
-        reportStoragePrefix + stableKey(from: userId)
-    }
-
-    private func signatureKey(for userId: String) -> String {
-        acknowledgedSignaturePrefix + stableKey(from: userId)
-    }
-
-    private func signatureForSnapshot(
-        sessionIds: [String],
-        pointCount: Int,
-        totalAreaM2: Double,
-        totalDurationSec: Double
-    ) -> String {
-        let payload = sessionIds.joined(separator: "|")
-        + "|p:\(pointCount)"
-        + "|a:\(totalAreaM2)"
-        + "|t:\(totalDurationSec)"
-        return stableKey(from: payload)
-    }
-
-    private func stableKey(from raw: String) -> String {
-        let digest = SHA256.hash(data: Data(raw.utf8))
-        return digest.map { String(format: "%02x", $0) }.joined()
-    }
-}
-
-@MainActor
-final class AuthFlowCoordinator: ObservableObject {
-    @Published var shouldShowEntryChoice: Bool = false
-    @Published var shouldShowSignIn: Bool = false
-    @Published var pendingUpgradeRequest: MemberUpgradeRequest? = nil
-    @Published var pendingGuestDataUpgradePrompt: GuestDataUpgradePrompt? = nil
-    @Published var guestDataUpgradeInProgress: Bool = false
-    @Published var guestDataUpgradeResult: GuestDataUpgradeReport? = nil
-    @Published private(set) var sessionStateSnapshot: AppSessionState = AppFeatureGate.currentSession()
-
-    private let guestModeKey = "auth.guest_mode.v1"
-    private let entryChoiceCompletedKey = "auth.entry_choice_completed.v1"
-    private let guestDataUpgradeService = GuestDataUpgradeService.shared
-    private let authSessionStore: AuthSessionStoreProtocol
-    private let profileStore: ProfileStoring
-    private let petSelectionStore: PetSelectionStoring
-    private let walkSessionMetadataStore: WalkSessionMetadataStore
-    private var onAuthenticated: (() -> Void)?
-    private var authSessionObserver: AnyCancellable?
-
-    /// 인증/프로필/선호 스토어 의존성을 주입해 인증 플로우 코디네이터를 초기화합니다.
-    /// - Parameters:
-    ///   - authSessionStore: 로그인 세션(토큰/사용자 식별자) 저장소입니다.
-    ///   - profileStore: 로컬 프로필 스냅샷 저장소입니다.
-    ///   - petSelectionStore: 반려견 선택 상태 저장소입니다.
-    ///   - walkSessionMetadataStore: 산책 메타데이터/선호 설정 저장소입니다.
-    init(
-        authSessionStore: AuthSessionStoreProtocol = DefaultAuthSessionStore.shared,
-        profileStore: ProfileStoring = ProfileStore.shared,
-        petSelectionStore: PetSelectionStoring = PetSelectionStore.shared,
-        walkSessionMetadataStore: WalkSessionMetadataStore = .shared
-    ) {
-        self.authSessionStore = authSessionStore
-        self.profileStore = profileStore
-        self.petSelectionStore = petSelectionStore
-        self.walkSessionMetadataStore = walkSessionMetadataStore
-        bindAuthSessionSync()
-    }
-
-    deinit {
-        authSessionObserver?.cancel()
-    }
-
-    var sessionState: AppSessionState {
-        sessionStateSnapshot
-    }
-
-    var isLoggedIn: Bool {
-        sessionState.isMember
-    }
-
-    var isGuestMode: Bool {
-        isLoggedIn == false && UserDefaults.standard.bool(forKey: guestModeKey)
-    }
-
-    func refresh() {
-        syncSessionStateSnapshot()
-        if isLoggedIn {
-            UserDefaults.standard.set(false, forKey: guestModeKey)
-            UserDefaults.standard.set(true, forKey: entryChoiceCompletedKey)
-            shouldShowEntryChoice = false
-            shouldShowSignIn = false
-            pendingUpgradeRequest = nil
-            return
-        }
-        pendingGuestDataUpgradePrompt = nil
-        guestDataUpgradeInProgress = false
-        guestDataUpgradeResult = nil
-        let didChooseEntryPath = UserDefaults.standard.bool(forKey: entryChoiceCompletedKey)
-        shouldShowEntryChoice = !didChooseEntryPath
-    }
-
-    func continueAsGuest() {
-        UserDefaults.standard.set(true, forKey: guestModeKey)
-        UserDefaults.standard.set(true, forKey: entryChoiceCompletedKey)
-        shouldShowEntryChoice = false
-    }
-
-    func chooseSignInFromEntry() {
-        UserDefaults.standard.set(true, forKey: entryChoiceCompletedKey)
-        shouldShowEntryChoice = false
-        shouldShowSignIn = true
-    }
-
-    func canAccess(_ feature: FeatureCapability) -> Bool {
-        AppFeatureGate.isAllowed(feature, session: sessionState)
-    }
-
-    @discardableResult
-    func requestAccess(feature: FeatureCapability, onAllowed: (() -> Void)? = nil) -> Bool {
-        let decision = AppFeatureGate.decision(for: feature, session: sessionState)
-        switch decision {
-        case .allowed:
-            onAllowed?()
-            return true
-        case .requiresMember(let trigger):
-            return requireMember(trigger: trigger, onAuthenticated: onAllowed)
-        }
-    }
-
-    @discardableResult
-    func requireMember(trigger: MemberUpgradeTrigger, onAuthenticated: (() -> Void)? = nil) -> Bool {
-        if isLoggedIn {
-            onAuthenticated?()
-            return true
-        }
-        self.onAuthenticated = onAuthenticated
-        pendingUpgradeRequest = MemberUpgradeRequest(trigger: trigger)
-        return false
-    }
-
-    func proceedToSignIn() {
-        pendingUpgradeRequest = nil
-        shouldShowSignIn = true
-    }
-
-    func dismissUpgradeRequest() {
-        pendingUpgradeRequest = nil
-        onAuthenticated = nil
-    }
-
-    func dismissSignIn() {
-        shouldShowSignIn = false
-        if isLoggedIn == false {
-            UserDefaults.standard.set(true, forKey: guestModeKey)
-        }
-        onAuthenticated = nil
-    }
-
-    func startReauthenticationFlow() {
-        authSessionStore.clearTokenSession()
-        syncSessionStateSnapshot()
-        pendingUpgradeRequest = nil
-        pendingGuestDataUpgradePrompt = nil
-        shouldShowEntryChoice = false
-        shouldShowSignIn = true
-        onAuthenticated = nil
-    }
-
-    func dismissGuestDataUpgradePrompt() {
-        pendingGuestDataUpgradePrompt = nil
-    }
-
-    func clearGuestDataUpgradeResult() {
-        guestDataUpgradeResult = nil
-    }
-
-    /// 현재 로그인 세션과 로컬 프로필 상태를 모두 정리하고 게스트 진입 상태로 전환합니다.
-    func signOut() {
-        authSessionStore.clear()
-        profileStore.removeAll()
-        petSelectionStore.clearSelectionState()
-        walkSessionMetadataStore.clearPreferences()
-        UserDefaults.standard.set(false, forKey: guestModeKey)
-        UserDefaults.standard.set(false, forKey: entryChoiceCompletedKey)
-        pendingUpgradeRequest = nil
-        pendingGuestDataUpgradePrompt = nil
-        guestDataUpgradeInProgress = false
-        guestDataUpgradeResult = nil
-        onAuthenticated = nil
-        refresh()
-    }
-
-    func startGuestDataUpgrade(forceRetry: Bool = false) {
-        guard let userId = currentMemberUserId() else {
-            #if DEBUG
-            print("[AuthFlow] startGuestDataUpgrade aborted: missing member session/token")
-            #endif
-            return
-        }
-        #if DEBUG
-        print("[AuthFlow] startGuestDataUpgrade user=\(userId) forceRetry=\(forceRetry)")
-        #endif
-        pendingGuestDataUpgradePrompt = nil
-        guestDataUpgradeInProgress = true
-        Task {
-            let report = await guestDataUpgradeService.runUpgrade(for: userId, forceRetry: forceRetry)
-            await MainActor.run {
-                self.guestDataUpgradeInProgress = false
-                self.guestDataUpgradeResult = report
-                #if DEBUG
-                if let report {
-                    print(
-                        "[AuthFlow] guest upgrade completed outstanding=\(report.hasOutstandingWork) pending=\(report.pendingCount) permanent=\(report.permanentFailureCount) lastError=\(report.lastErrorCode ?? "none")"
-                    )
-                } else {
-                    print("[AuthFlow] guest upgrade completed with nil report")
-                }
-                #endif
-            }
-        }
-    }
-
-    func latestGuestDataUpgradeReport() -> GuestDataUpgradeReport? {
-        guard let userId = currentMemberUserId() else {
-            return nil
-        }
-        return guestDataUpgradeService.latestReport(for: userId)
-    }
-
-    func completeSignIn() {
-        syncSessionStateSnapshot()
-        UserDefaults.standard.set(false, forKey: guestModeKey)
-        UserDefaults.standard.set(true, forKey: entryChoiceCompletedKey)
-        shouldShowSignIn = false
-        shouldShowEntryChoice = false
-        pendingUpgradeRequest = nil
-        if let userId = currentMemberUserId() {
-            pendingGuestDataUpgradePrompt = guestDataUpgradeService.pendingPrompt(for: userId)
-            guestDataUpgradeResult = guestDataUpgradeService.latestReport(for: userId)
-        }
-        let completion = onAuthenticated
-        onAuthenticated = nil
-        completion?()
-    }
-
-    /// 현재 인증 세션/프로필 스토어에서 사용자 식별자를 조회합니다.
-    /// - Returns: 로그인 사용자 ID가 있으면 반환하고, 없으면 `nil`을 반환합니다.
-    private func currentMemberUserId() -> String? {
-        guard let sessionUserId = authSessionStore.currentIdentity()?.userId,
-              sessionUserId.isEmpty == false else {
-            return nil
-        }
-        guard authSessionStore.currentTokenSession() != nil else {
-            return nil
-        }
-        return sessionUserId
-    }
-
-    /// 저장소 기준 최신 세션 상태를 계산해 `@Published` 스냅샷으로 반영합니다.
-    /// - Returns: 없음. 세션 전환이 있으면 SwiftUI 갱신 트리거를 발생시킵니다.
-    private func syncSessionStateSnapshot() {
-        sessionStateSnapshot = AppFeatureGate.currentSession()
-    }
-
-    /// 인증 세션 변경 알림을 구독해 코디네이터 상태를 즉시 동기화합니다.
-    private func bindAuthSessionSync() {
-        authSessionObserver = NotificationCenter.default.publisher(for: .authSessionDidChange)
-            .receive(on: RunLoop.main)
-            .sink { [weak self] _ in
-                self?.refresh()
-            }
-    }
-}
-
-struct MemberUpgradeSheetView: View {
-    let request: MemberUpgradeRequest
-    let onUpgrade: () -> Void
-    let onLater: () -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 14) {
-            Text(request.trigger.title)
-                .font(.appFont(for: .Bold, size: 22))
-                .foregroundStyle(Color.appTextDarkGray)
-            Text(request.trigger.message)
-                .font(.appFont(for: .Regular, size: 14))
-                .foregroundStyle(Color.appTextDarkGray)
-            VStack(alignment: .leading, spacing: 8) {
-                Text("• 계정 연동 후 자동 백업")
-                Text("• 기기 변경 시 기록 복원")
-                Text("• 로그인 완료 후 현재 화면으로 복귀")
-            }
-            .font(.appFont(for: .Regular, size: 13))
-            .foregroundStyle(Color.appTextDarkGray)
-            HStack(spacing: 10) {
-                Button("나중에") {
-                    onLater()
-                }
-                .accessibilityIdentifier("sheet.memberUpgrade.later")
-                .frame(maxWidth: .infinity)
-                .padding(.vertical, 12)
-                .background(Color.appYellowPale)
-                .foregroundStyle(Color.appTextDarkGray)
-                .cornerRadius(10)
-
-                Button("로그인하고 계속") {
-                    onUpgrade()
-                }
-                .accessibilityIdentifier("sheet.memberUpgrade.signin")
-                .frame(maxWidth: .infinity)
-                .padding(.vertical, 12)
-                .background(Color.appGreen)
-                .foregroundStyle(Color.white)
-                .cornerRadius(10)
-            }
-        }
-        .padding(20)
-        .background(Color.white)
-    }
-}
-
-struct GuestDataUpgradePromptSheetView: View {
-    let prompt: GuestDataUpgradePrompt
-    let onImport: () -> Void
-    let onLater: () -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 14) {
-            Text(prompt.shouldEmphasizeRetry ? "산책 데이터 이관 재시도" : "게스트 산책 데이터 가져오기")
-                .font(.appFont(for: .Bold, size: 22))
-                .foregroundStyle(Color.appTextDarkGray)
-            Text("로그인 전에 기록한 산책 데이터를 계정으로 이관합니다. 중복 없이 안전하게 처리돼요.")
-                .font(.appFont(for: .Regular, size: 14))
-                .foregroundStyle(Color.appTextDarkGray)
-            VStack(alignment: .leading, spacing: 6) {
-                Text("세션 \(prompt.snapshot.sessionCount)건")
-                Text("포인트 \(prompt.snapshot.pointCount)건")
-                Text("누적 면적 \(prompt.snapshot.totalAreaM2.calculatedAreaString)")
-                Text("누적 시간 \(prompt.snapshot.totalDurationSec.walkingTimeInterval)")
-            }
-            .font(.appFont(for: .Regular, size: 13))
-            .foregroundStyle(Color.appTextDarkGray)
-            HStack(spacing: 10) {
-                Button("나중에") {
-                    onLater()
-                }
-                .frame(maxWidth: .infinity)
-                .padding(.vertical, 12)
-                .background(Color.appYellowPale)
-                .foregroundStyle(Color.appTextDarkGray)
-                .cornerRadius(10)
-
-                Button(prompt.shouldEmphasizeRetry ? "다시 가져오기" : "가져오기") {
-                    onImport()
-                }
-                .frame(maxWidth: .infinity)
-                .padding(.vertical, 12)
-                .background(Color.appGreen)
-                .foregroundStyle(Color.white)
-                .cornerRadius(10)
-            }
-        }
-        .padding(20)
-        .background(Color.white)
-    }
-}
