+++ dogArea/Views/SigningView/Components/AuthUserInfo.swift
@@ -1,0 +1,12 @@
+//
+//  AuthUserInfo.swift
+//  dogArea
+//
+
+import Foundation
+
+struct AuthUserInfo: Identifiable, Hashable, TimeCheckable {
+    var createdAt: TimeInterval
+    let id: String
+    let name: String?
+}
