+++ scripts/auth_signup_validation_unit_check.swift
@@ -15,7 +15,15 @@ func load(_ relativePath: String) -> String {
     return String(decoding: data, as: UTF8.self)
 }
 
-let signInView = load("dogArea/Views/SigningView/SignInView.swift")
+func loadMany(_ relativePaths: [String]) -> String {
+    relativePaths.map(load).joined(separator: "\n")
+}
+
+let signInView = loadMany([
+    "dogArea/Views/SigningView/SignInView.swift",
+    "dogArea/Views/SigningView/Components/AuthUserInfo.swift",
+    "dogArea/Views/SigningView/Components/EmailSignUpSheetView.swift"
+])
 let infra = load("dogArea/Source/Infrastructure/Supabase/SupabaseInfrastructure.swift")
 
 assertTrue(
