+++ scripts/map_home_viewmodel_boundary_unit_check.swift
@@ -16,6 +16,10 @@ func load(_ relativePath: String) -> String {
     return String(decoding: data, as: UTF8.self)
 }
 
+func loadMany(_ relativePaths: [String]) -> String {
+    relativePaths.map(load).joined(separator: "\n")
+}
+
 func extractTypeBody(source: String, declarationPrefix: String) -> String? {
     guard let declarationRange = source.range(of: declarationPrefix) else { return nil }
     guard let openBrace = source[declarationRange.lowerBound...].firstIndex(of: "{") else { return nil }
@@ -38,7 +42,12 @@ func extractTypeBody(source: String, declarationPrefix: String) -> String? {
 }
 
 let mapViewModel = load("dogArea/Views/MapView/MapViewModel.swift")
-let homeViewModel = load("dogArea/Views/HomeView/HomeViewModel.swift")
+let homeViewModel = loadMany([
+    "dogArea/Views/HomeView/HomeViewModel.swift",
+    "dogArea/Source/Domain/Home/Models/HomeMissionModels.swift",
+    "dogArea/Source/Domain/Home/Stores/IndoorMissionStore.swift",
+    "dogArea/Source/Domain/Home/Stores/SeasonMotionStore.swift"
+])
 
 let mapBody = extractTypeBody(source: mapViewModel, declarationPrefix: "class MapViewModel")
 let homeBody = extractTypeBody(source: homeViewModel, declarationPrefix: "final class HomeViewModel")
