+++ scripts/rival_observability_metrics_unit_check.swift
@@ -23,7 +23,16 @@ func load(_ relativePath: String) -> String {
     return String(decoding: data, as: UTF8.self)
 }
 
-let defaults = load("dogArea/Source/UserdefaultSetting.swift")
+func loadMany(_ relativePaths: [String]) -> String {
+    relativePaths.map(load).joined(separator: "\n")
+}
+
+let defaults = loadMany([
+    "dogArea/Source/UserdefaultSetting.swift",
+    "dogArea/Source/AppSession/AppFeatureGate.swift",
+    "dogArea/Source/AppSession/GuestDataUpgradeService.swift",
+    "dogArea/Source/AppSession/AuthFlowCoordinator.swift"
+])
 let rivalViewModel = load("dogArea/Views/ProfileSettingView/RivalTabViewModel.swift")
 let rivalSpec = load("docs/rival-tab-ux-usecase-spec-v1.md")
 let gameLayerSpec = load("docs/game-layer-observability-qa-v1.md")
