+++ scripts/sync_walk_404_policy_unit_check.swift
@@ -15,8 +15,17 @@ func load(_ relativePath: String) -> String {
     return String(decoding: data, as: UTF8.self)
 }
 
+func loadMany(_ relativePaths: [String]) -> String {
+    relativePaths.map(load).joined(separator: "\n")
+}
+
 let infra = load("dogArea/Source/Infrastructure/Supabase/SupabaseInfrastructure.swift")
-let syncStore = load("dogArea/Source/UserdefaultSetting.swift")
+let syncStore = loadMany([
+    "dogArea/Source/UserdefaultSetting.swift",
+    "dogArea/Source/AppSession/AppFeatureGate.swift",
+    "dogArea/Source/AppSession/GuestDataUpgradeService.swift",
+    "dogArea/Source/AppSession/AuthFlowCoordinator.swift"
+])
 let mapVM = load("dogArea/Views/MapView/MapViewModel.swift")
 let homeView = load("dogArea/Views/HomeView/HomeView.swift")
