+++ scripts/walk_session_pet_canonicalization_unit_check.swift
@@ -15,11 +15,20 @@ func load(_ relativePath: String) -> String {
     return String(decoding: data, as: UTF8.self)
 }
 
+func loadMany(_ relativePaths: [String]) -> String {
+    relativePaths.map(load).joined(separator: "\n")
+}
+
 let walkRepository = load("dogArea/Source/Data/Walk/WalkRepository.swift")
 let walkBackfill = load("dogArea/Source/Data/Walk/WalkBackfillDTO.swift")
 let mapModel = load("dogArea/Source/Domain/Map/Models/MapModel.swift")
 let mapViewModel = load("dogArea/Views/MapView/MapViewModel.swift")
-let homeViewModel = load("dogArea/Views/HomeView/HomeViewModel.swift")
+let homeViewModel = loadMany([
+    "dogArea/Views/HomeView/HomeViewModel.swift",
+    "dogArea/Source/Domain/Home/Models/HomeMissionModels.swift",
+    "dogArea/Source/Domain/Home/Stores/IndoorMissionStore.swift",
+    "dogArea/Source/Domain/Home/Stores/SeasonMotionStore.swift"
+])
 let walkListModel = load("dogArea/Views/WalkListView/WalkListModel.swift")
 let walkListViewModel = load("dogArea/Views/WalkListView/WalkListViewModel.swift")
