+++ scripts/weather_ux_stage3_unit_check.swift
@@ -15,8 +15,17 @@ func load(_ relativePath: String) -> String {
     return String(decoding: data, as: UTF8.self)
 }
 
+func loadMany(_ relativePaths: [String]) -> String {
+    relativePaths.map(load).joined(separator: "\n")
+}
+
 let homeView = load("dogArea/Views/HomeView/HomeView.swift")
-let homeViewModel = load("dogArea/Views/HomeView/HomeViewModel.swift")
+let homeViewModel = loadMany([
+    "dogArea/Views/HomeView/HomeViewModel.swift",
+    "dogArea/Source/Domain/Home/Models/HomeMissionModels.swift",
+    "dogArea/Source/Domain/Home/Stores/IndoorMissionStore.swift",
+    "dogArea/Source/Domain/Home/Stores/SeasonMotionStore.swift"
+])
 let mapView = load("dogArea/Views/MapView/MapView.swift")
 let mapViewModel = load("dogArea/Views/MapView/MapViewModel.swift")
 let spec = load("docs/weather-ux-fallback-accessibility-v1.md")
